+++ components/Container/Container.module.scss
@@ -3,17 +3,21 @@
     margin-right: auto;
     padding-left: 15px;
     padding-right: 15px;
-    // outline: 1px solid blue;
-    
+  
     @media screen and (min-width: 480px) {
-        width: 480px;
+      width: 480px;
     }
-
+  
+    // @media screen and (max-width: 767px) {
+    //   overflow-x: hidden;
+    // }
+  
     @media screen and (min-width: 768px) {
-        width: 768px;
+      width: 768px;
     }
-
-    @media screen and (min-width: 1024px) {
-        width: 1024px;
+  
+    @media screen and (min-width: 1200px) {
+      width: 1200px;
     }
-}
+  }
+
