+++ components/Header/Logo/Logo.module.scss
@@ -1,0 +1,32 @@
+%logoTemplate {
+    display: flex;
+    padding: 10px;
+    align-items: center;
+    font-weight: 700;
+    color: white;
+    font-style: italic;
+    font-size: 12px;
+    & svg {
+      fill: var(--accentColor);
+      width: 80px;
+      height: 70px;
+    }
+    @media screen and (max-width: 768px) {
+      font-size: 8px;
+  
+      & svg {
+        width: 50px;
+        height: 30px;
+      }
+    }
+  }
+  .logo {
+    @extend %logoTemplate;
+  }
+  
+  .pageLogo {
+    @extend %logoTemplate;
+  
+    color: black;
+  }
+
