+++ components/Section/Section.module.scss
@@ -1,4 +1,5 @@
 .section {
     padding-top: 70px;
     padding-bottom: 70px;
-}
+  }
+
